Created BalanceController, BalanceRepository, BalanceService, BalanceResponse and BalanceRequest. Also created two interfaces to repository and service.

+++ Backend/DTO/UserDTO/UserResponse.cs
@@ -12,9 +12,9 @@ namespace Backend.DTO.UserDTO
         public DateOnly BirthDate { get; set; }
         public int? PhoneNumber { get; set; }
         public bool NewsLetterIsSubscribed { get; set; }
-        public int Balance { get; set; }
-        public int Profit { get; set; }
-        public int Loss { get; set; }
+        public decimal Balance { get; set; }
+        public decimal Profit { get; set; }
+        public decimal Loss { get; set; }
         public Role Role { get; set; }
         public DateTime? ExcludedUntil { get; set; }
     }
